api/requirements.txt
fastapi
uvicorn
redis>=5.0.1
sse-starlette>=1.8.2
python-dotenv>=1.0.0
pydantic>=2.0.0 